Marked feature as disabled
