+++ src/data/flag/resolver.js
@@ -1,22 +1,11 @@
 import ApollosConfig from '@apollosproject/config'
 import { get } from 'lodash'
+import { currentUserCanUseFeature } from '.'
 
 const resolver = {
     Query: {
-        flagStatus: async (root, { key }, { dataSources }) => {
-            const flag = get(ApollosConfig, `FEATURE_FLAGS.${key}`, null)
-            if (flag && flag.status === "LIVE") {
-                if (flag.securityGroupId) {
-                    return await dataSources.Auth.isInSecurityGroup(flag.securityGroupId)
-                        ? "LIVE"
-                        : "DISABLED"
-                }
-
-                return flag.status
-            }
-
-            return "DISABLED"
-        },
+        flagStatus: (root, { key }, { dataSources }) =>
+            currentUserCanUseFeature(key)
     }
 }
